+++ examples/originate.py
@@ -4,6 +4,8 @@ import asyncio
 from panoramisk import Manager
 
 
+# logging.basicConfig(level=logging.DEBUG)
+
 loop = asyncio.get_event_loop()
 
 @asyncio.coroutine