+++ docs/conftest.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+import pytest
+
+
+@pytest.fixture
+def inject_file():
+    return __file__
+